requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# File processing
pdfminer.six==20221105
python-docx==1.1.0
docx2txt==0.8

# AI/ML - Core NLP and ML libraries
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
huggingface-hub==0.16.4

# Visualization
plotly==5.17.0

# Configuration
pyyaml==6.0.1

# HTTP requests
requests==2.31.0

# Date/time handling
python-dateutil==2.8.2

# Database & Authentication
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing (optional for production)
# pytest==7.4.3
# pytest-cov==4.1.0

# Code quality (optional for production)
# black==23.11.0
# flake8==6.1.0